update delay function and delay tests
