+++ Tests/deferredTests/DeferredExtrasTests.swift
@@ -704,67 +704,3 @@ class DeferredExtrasTests: XCTestCase
     XCTAssertEqual(s4.0.value.map(String.init), s4.3.value)
   }
 }
-
-class DelayTests: XCTestCase
-{
-  func testDelayValue()
-  {
-    let t1 = 0.05
-    let d1 = Deferred(value: Date()).delay(seconds: t1).map { Date().timeIntervalSince($0) }
-    XCTAssert(d1.value! >= t1)
-  }
-
-  func testDelayError()
-  {
-    let d1 = Deferred(value: Date()).delay(until: .distantFuture)
-    let d2 = d1.delay(until: .distantFuture)
-    d1.cancel()
-    XCTAssert(d2.value == nil)
-  }
-
-  func testCancelDelay()
-  {
-    let d1 = Deferred(value: Date()).delay(until: .distantFuture)
-    let d2 = d1.delay(until: .distantFuture)
-    d2.cancel()
-    d1.cancel()
-    XCTAssert(d1.value == d2.value)
-  }
-
-  func testAbandonedDelay()
-  {
-    let d1 = Deferred(value: 1).delay(until: .distantFuture)
-    let d2 = d1.delay(.milliseconds(500)).map(transform: { 2*$0 })
-    d2.cancel()
-    d1.cancel()
-    XCTAssert(d2.result.isError)
-    XCTAssert(d1.result.isError)
-  }
-
-  func testSourceSlowerThanDelay()
-  {
-    let d1 = Deferred(value: nzRandom()).delay(.milliseconds(100))
-    let d2 = d1.delay(until: .now() + .microseconds(100))
-    XCTAssert(d1.result.value == d2.result.value)
-    XCTAssert(d2.result.isValue)
-  }
-
-  func testDistantFuture()
-  { // corelibs-foundation formerly required a special-case for .distantFuture
-    // (see https://bugs.swift.org/browse/SR-5706)
-    let time = DispatchTime.distantFuture
-    XCTAssert(time > .now(), "please use Swift 4.0.2 or later on Linux; see https://bugs.swift.org/browse/SR-5706")
-  }
-
-  func testDistantFutureDelay()
-  {
-    let d1 = Deferred(value: Date())
-    let d2 = d1.delay(until: .distantFuture)
-
-    let e1 = expectation(description: "immediate")
-    d1.onValue { _ in e1.fulfill() }
-    waitForExpectations(timeout: 1.0)
-
-    d2.cancel()
-  }
-}
